% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clickrup-package.R
\docType{package}
\name{clickrup-package}
\alias{clickrup}
\alias{clickrup-package}
\title{clickrup: 'ClickUp' 'API' Client}
\description{
A client to interact wiht the 'ClickUp' 'API' \url{https://clickup.com/api/}. 'ClickUp' is a project management app, available at \url{https://clickup.com}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonthegeek/clickrup}
  \item \url{https://jonthegeek.github.io/clickrup/}
  \item Report bugs at \url{https://github.com/jonthegeek/clickrup/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
